feat（bom） 增加供货周期+ 零件购买计划优化

+++ src/main/java/com/olivia/peanut/aps/api/entity/apsSchedulingGoodsBom/ApsSchedulingGoodsBomDto.java
@@ -57,7 +57,7 @@ public class ApsSchedulingGoodsBomDto extends BaseEntityDto {
    *  使用量
    */
   @NotNull(message = "使用量不能为空", groups = {InsertCheck.class, UpdateCheck.class})
-  private Integer bomUsage;
+  private BigDecimal bomUsage;
   /***
    *  规格
    */
