+++ src/main/java/com/olivia/peanut/aps/model/ApsSchedulingGoodsBom.java
@@ -52,7 +52,7 @@ public class ApsSchedulingGoodsBom extends BaseEntity {
   /***
    *  使用量
    */
-  private Integer bomUsage;
+  private BigDecimal bomUsage;
   /***
    *  规格
    */
